Change the bezier curve selection tool options widget to use KisAngleSelector

+++ libs/basicflakes/tools/KoCreatePathTool.h
@@ -111,7 +111,7 @@ protected:
 
 private:
     Q_DECLARE_PRIVATE(KoCreatePathTool)
-    Q_PRIVATE_SLOT(d_func(), void angleDeltaChanged(int))
+    Q_PRIVATE_SLOT(d_func(), void angleDeltaChanged(qreal))
     Q_PRIVATE_SLOT(d_func(), void angleSnapChanged(int))
     Q_PRIVATE_SLOT(d_func(), void autoSmoothCurvesChanged(bool))
 };
